implement BRLitecoinVerifyDifficulty() and add litecoin addrParams
